fix: property tests fast-check constraints

+++ backend/tests/property/productPricing.property.test.ts
@@ -4,9 +4,9 @@ describe("product pricing invariants", () => {
   it("final price invariants", () => {
     fc.assert(
       fc.property(
-        fc.float({ min: 0.01, max: 1_000_000, noNaN: true }),
-        fc.float({ min: 0, max: 1, noNaN: true }),
-        fc.float({ min: 0, max: 0.28, noNaN: true }),
+        fc.double({ min: 0.01, max: 1_000_000, noNaN: true }),
+        fc.double({ min: 0, max: 1, noNaN: true }),
+        fc.double({ min: 0, max: 0.28, noNaN: true }),
         (price, discountRate, gstRate) => {
           const discount = price * discountRate;
           const gst = price * gstRate;
